scrimlet-reconcilers: move status types to bootstrap-agent-lockstep-types
