+++ sled-agent/bootstrap-agent-lockstep-types/Cargo.toml
@@ -9,6 +9,7 @@ workspace = true
 
 [dependencies]
 anyhow.workspace = true
+chrono.workspace = true
 iddqd.workspace = true
 omicron-common.workspace = true
 omicron-passwords.workspace = true
@@ -17,6 +18,7 @@ schemars.workspace = true
 serde.workspace = true
 sled-agent-types.workspace = true
 sled-hardware-types.workspace = true
+slog.workspace = true
 strum.workspace = true
 wicketd-commission-types.workspace = true
