+++ sled-agent/bootstrap-agent-lockstep-types/src/lib.rs
@@ -36,6 +36,8 @@ use strum::IntoEnumIterator;
 pub use wicketd_commission_types::rack_setup::ServiceIpPoolConfig;
 pub use wicketd_commission_types::rack_setup::ServiceIpPoolError;
 
+pub mod scrimlet_reconcilers;
+
 /// Configuration for the "rack setup service".
 ///
 /// The Rack Setup Service should be responsible for one-time setup actions,
