+++ sled-agent/scrimlet-reconcilers/src/dpd_reconciler/nat.rs
@@ -6,6 +6,10 @@
 //!
 //! Does not modify non-service NAT entries.
 
+use bootstrap_agent_lockstep_types::scrimlet_reconcilers::dpd::{
+    DpdNatReconcilerStatus, DpdNatReconcilerStatusNatEntry,
+    DpdNatReconcilerStatusNatEntryFailure,
+};
 use daft::Diffable;
 use dpd_client::Client;
 use futures::Stream;
@@ -30,126 +34,6 @@ type DpdClientError = dpd_client::Error<dpd_client::types::Error>;
 
 const SINGLE_REQUEST_LIMIT: Option<NonZeroU32> =
     Some(NonZeroU32::new(128).unwrap());
-
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
-pub struct DpdNatReconcilerStatusNatEntry {
-    pub external_ip: IpAddr,
-    pub first_port: u16,
-    pub last_port: u16,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
-pub struct DpdNatReconcilerStatusNatEntryFailure {
-    pub entry: DpdNatReconcilerStatusNatEntry,
-    pub error: String,
-}
-
-/// Status of reconciling service zone NAT entries with `dpd`.
-#[derive(Debug, Clone)]
-pub enum DpdNatReconcilerStatus {
-    /// Reconciliation was skipped because the bootstore contains no NAT entry
-    /// config information.
-    NoNatEntriesConfig,
-
-    /// Reconciliation failed while attempting to read the current set of
-    /// entries from `dpd`.
-    FailedReadingCurrentDpdNatEntries(String),
-
-    /// Reconciliation failed because the bootstore config contained an illegal
-    /// combination of entries (e.g., two zones with identical NAT entries).
-    InvalidSystemNetworkingConfig(String),
-
-    /// Reconciliation completed successfully.
-    Success {
-        /// Set of zone IDs whose NAT entries were already correct in `dpd` and
-        /// left unchanged.
-        unchanged: BTreeSet<OmicronZoneUuid>,
-
-        /// List of NAT entries removed.
-        removed: Vec<DpdNatReconcilerStatusNatEntry>,
-
-        /// Map of zone NAT entries created.
-        created: BTreeMap<OmicronZoneUuid, DpdNatReconcilerStatusNatEntry>,
-    },
-
-    /// Reconciliation completed but had at least one failure.
-    PartialSuccess {
-        /// Set of zone IDs whose NAT entries were already correct in `dpd` and
-        /// left unchanged.
-        unchanged: BTreeSet<OmicronZoneUuid>,
-
-        /// List of NAT entries successfully removed.
-        removed: Vec<DpdNatReconcilerStatusNatEntry>,
-
-        /// List of NAT entries we tried but failed to remove.
-        remove_failures: Vec<DpdNatReconcilerStatusNatEntryFailure>,
-
-        /// Map of zone NAT entries successfully created.
-        created: BTreeMap<OmicronZoneUuid, DpdNatReconcilerStatusNatEntry>,
-
-        /// Map of zone NAT entries we tried but failed to create.
-        create_failures:
-            BTreeMap<OmicronZoneUuid, DpdNatReconcilerStatusNatEntryFailure>,
-    },
-}
-
-impl slog::KV for DpdNatReconcilerStatus {
-    fn serialize(
-        &self,
-        _record: &slog::Record<'_>,
-        serializer: &mut dyn slog::Serializer,
-    ) -> slog::Result {
-        match self {
-            DpdNatReconcilerStatus::NoNatEntriesConfig => serializer.emit_str(
-                "nat-reconciler-skipped".into(),
-                "no NAT entries present in config",
-            ),
-            DpdNatReconcilerStatus::FailedReadingCurrentDpdNatEntries(
-                reason,
-            ) => serializer.emit_str("nat-reconciler-failed".into(), reason),
-            DpdNatReconcilerStatus::InvalidSystemNetworkingConfig(reason) => {
-                serializer.emit_arguments(
-                    "nat-reconciler-failed".into(),
-                    &format_args!("invalid system networking config: {reason}"),
-                )
-            }
-            DpdNatReconcilerStatus::Success { unchanged, removed, created } => {
-                // Only show a summary count; we have individual log statements
-                // for each create/remove.
-                for (key, val) in [
-                    ("nat-entries-unchanged", unchanged.len()),
-                    ("nat-entries-successfully-removed", removed.len()),
-                    ("nat-entries-failed-to-remove", 0),
-                    ("nat-entries-successfully-created", created.len()),
-                    ("nat-entries-failed-to-create", 0),
-                ] {
-                    serializer.emit_usize(key.into(), val)?;
-                }
-                Ok(())
-            }
-            DpdNatReconcilerStatus::PartialSuccess {
-                unchanged,
-                removed,
-                remove_failures,
-                created,
-                create_failures,
-            } => {
-                // Only show a summary count; we have individual log statements
-                // for each create/remove.
-                for (key, val) in [
-                    ("nat-entries-unchanged", unchanged.len()),
-                    ("nat-entries-successfully-removed", removed.len()),
-                    ("nat-entries-failed-to-remove", remove_failures.len()),
-                    ("nat-entries-successfully-created", created.len()),
-                    ("nat-entries-failed-to-create", create_failures.len()),
-                ] {
-                    serializer.emit_usize(key.into(), val)?;
-                }
-                Ok(())
-            }
-        }
-    }
-}
 
 /// Perform reconciliation.
 ///
