+++ sled-agent/scrimlet-reconcilers/src/dpd_reconciler/port_reconciler/tests.rs
@@ -137,7 +137,7 @@ fn plan_all_unchanged() {
     )
     .expect("plan should succeed");
 
-    assert_eq!(plan.unchanged, BTreeSet::from([qsfp0]));
+    assert_eq!(plan.unchanged, BTreeSet::from([qsfp0.to_string()]));
     assert!(plan.to_clear.is_empty());
     assert!(plan.to_apply.is_empty());
 
@@ -320,7 +320,7 @@ fn plan_mix() {
     .expect("plan should succeed");
 
     // qsfp0: unchanged
-    assert_eq!(plan.unchanged, BTreeSet::from([qsfp0]));
+    assert_eq!(plan.unchanged, BTreeSet::from([qsfp0.to_string()]));
     // qsfp2: in dpd but not desired
     assert_eq!(plan.to_clear, BTreeSet::from([qsfp2]));
     // qsfp1: changed, qsfp3: new
@@ -420,7 +420,7 @@ fn plan_link_local_addrs_ignored_from_dpd() {
     )
     .expect("plan should succeed");
 
-    assert_eq!(plan.unchanged, BTreeSet::from([qsfp0]));
+    assert_eq!(plan.unchanged, BTreeSet::from([qsfp0.to_string()]));
     assert!(plan.to_clear.is_empty());
     assert!(plan.to_apply.is_empty());
 
@@ -687,7 +687,7 @@ impl TestInput {
 
     // Build the set of expected unchanged port names based on our arbitrary
     // inputs.
-    fn expected_unchanged(&self) -> BTreeSet<DpdQsfp> {
+    fn expected_unchanged(&self) -> BTreeSet<String> {
         self.ports
             .iter()
             .filter_map(|(port_id, input)| match input {
@@ -695,7 +695,7 @@ impl TestInput {
                 | SwitchPortSettingsTestInput::DesiredSwitch0(_)
                 | SwitchPortSettingsTestInput::DesiredSwitch1(_) => None,
                 SwitchPortSettingsTestInput::DpdAndSwitch0Same(_) => {
-                    Some(port_id.to_dpd())
+                    Some(port_id.to_dpd().to_string())
                 }
                 SwitchPortSettingsTestInput::DpdAndSwitch0Changed {
                     dpd,
@@ -707,7 +707,7 @@ impl TestInput {
                     if dpd.to_dpd_settings(port_id)
                         == switch0.to_dpd_settings(port_id)
                     {
-                        Some(port_id.to_dpd())
+                        Some(port_id.to_dpd().to_string())
                     } else {
                         None
                     }
@@ -941,13 +941,17 @@ async fn proptest_full_reconciliation() {
                 );
                 assert_eq!(
                     cleared,
-                    input.expected_to_clear(),
+                    input
+                        .expected_to_clear()
+                        .into_iter()
+                        .map(|p| p.to_string())
+                        .collect::<BTreeSet<_>>(),
                     "incorrect cleared"
                 );
                 let expected_applied = input
                     .expected_to_apply()
                     .keys()
-                    .cloned()
+                    .map(|p| p.to_string())
                     .collect::<BTreeSet<_>>();
                 assert_eq!(applied, expected_applied, "incorrect applied");
