+++ sled-agent/scrimlet-reconcilers/src/lldpd_reconciler.rs
@@ -13,6 +13,7 @@ use crate::reconciler_task::Reconciler;
 use crate::switch_zone_slot::ThisSledSwitchSlot;
 use anyhow::anyhow;
 use anyhow::bail;
+use bootstrap_agent_lockstep_types::scrimlet_reconcilers::lldpd::LldpdReconcilerStatus;
 use scuffle::AddPropertyGroupFlags;
 use scuffle::EditPropertyGroups;
 use scuffle::HasComposedPropertyGroups;
@@ -51,31 +52,6 @@ mod property_name {
     pub(super) const SYSTEM_NAME: &str = "system_name";
     pub(super) const SYSTEM_DESCRIPTION: &str = "system_description";
     pub(super) const MANAGEMENT_ADDRS: &str = "management_addrs";
-}
-
-#[derive(Debug, Clone)]
-pub enum LldpdReconcilerStatus {
-    Failed(String),
-    SkippedConfigUpToDate,
-    Reconciled { ports: BTreeMap<String, LldpAdminStatus> },
-}
-
-impl slog::KV for LldpdReconcilerStatus {
-    fn serialize(
-        &self,
-        _record: &slog::Record<'_>,
-        serializer: &mut dyn slog::Serializer,
-    ) -> slog::Result {
-        match self {
-            LldpdReconcilerStatus::Failed(reason) => {
-                serializer.emit_str("lldpd".into(), reason)
-            }
-            LldpdReconcilerStatus::SkippedConfigUpToDate => serializer
-                .emit_str("lldpd".into(), "skipped: config up-to-date"),
-            LldpdReconcilerStatus::Reconciled { ports } => serializer
-                .emit_usize("lldpd-reconciled-ports".into(), ports.len()),
-        }
-    }
 }
 
 #[derive(Debug)]
