+++ sled-agent/scrimlet-reconcilers/src/mgd_reconciler.rs
@@ -8,6 +8,7 @@
 use crate::ScrimletReconcilersMode;
 use crate::reconciler_task::Reconciler;
 use crate::switch_zone_slot::ThisSledSwitchSlot;
+use bootstrap_agent_lockstep_types::scrimlet_reconcilers::mgd::MgdReconcilerStatus;
 use mg_admin_client::Client;
 use sled_agent_types::system_networking::SystemNetworkingConfig;
 use slog::Logger;
@@ -16,33 +17,6 @@ use std::time::Duration;
 mod bfd_reconciler;
 mod bgp_reconciler;
 mod static_route_reconciler;
-
-pub use bfd_reconciler::MgdBfdOperationFailure;
-pub use bfd_reconciler::MgdBfdReconcilerStatus;
-pub use bgp_reconciler::MgdBgpReconcilerStatus;
-pub use bgp_reconciler::MgdBgpReconcilerStatusOpCount;
-pub use static_route_reconciler::MgdStaticRouteReconcilerStatus;
-
-#[derive(Debug, Clone)]
-pub struct MgdReconcilerStatus {
-    pub bfd_status: MgdBfdReconcilerStatus,
-    pub bgp_status: MgdBgpReconcilerStatus,
-    pub static_routes_status: MgdStaticRouteReconcilerStatus,
-}
-
-impl slog::KV for MgdReconcilerStatus {
-    fn serialize(
-        &self,
-        record: &slog::Record<'_>,
-        serializer: &mut dyn slog::Serializer,
-    ) -> slog::Result {
-        let Self { bfd_status, bgp_status, static_routes_status } = self;
-        bfd_status.serialize(record, serializer)?;
-        bgp_status.serialize(record, serializer)?;
-        static_routes_status.serialize(record, serializer)?;
-        Ok(())
-    }
-}
 
 #[derive(Debug)]
 pub(crate) struct MgdReconciler {
