+++ sled-agent/scrimlet-reconcilers/src/mgd_reconciler/bgp_reconciler.rs
@@ -7,6 +7,9 @@
 use crate::switch_zone_slot::ThisSledSwitchSlot;
 use anyhow::Context;
 use anyhow::bail;
+use bootstrap_agent_lockstep_types::scrimlet_reconcilers::mgd::{
+    MgdBgpReconcilerStatus, MgdBgpReconcilerStatusOpCount,
+};
 use daft::Diffable;
 use mg_admin_client::Client;
 use mg_admin_client::ResponseValue as MgdResponseValue;
@@ -51,164 +54,6 @@ type MgdClientError = mg_admin_client::Error<mg_admin_client::types::Error>;
 const BGP_SESSION_RESOLUTION: u64 = 100;
 
 const BGP_PORT: u16 = 179;
-
-#[derive(Debug, Clone, Default)]
-pub struct MgdBgpReconcilerStatusOpCount {
-    routers_deleted: usize,
-    routers_updated: usize,
-    routers_created: usize,
-
-    origin4_deleted: usize,
-    origin4_updated: usize,
-    origin4_created: usize,
-
-    origin6_deleted: usize,
-    origin6_updated: usize,
-    origin6_created: usize,
-
-    shapers_deleted: usize,
-    shapers_updated: usize,
-    shapers_created: usize,
-
-    checkers_deleted: usize,
-    checkers_updated: usize,
-    checkers_created: usize,
-
-    numbered_peers_deleted: usize,
-    numbered_peers_updated: usize,
-    numbered_peers_created: usize,
-
-    unnumbered_peers_deleted: usize,
-    unnumbered_peers_updated: usize,
-    unnumbered_peers_created: usize,
-}
-
-impl slog::KV for MgdBgpReconcilerStatusOpCount {
-    fn serialize(
-        &self,
-        _record: &slog::Record<'_>,
-        serializer: &mut dyn slog::Serializer,
-    ) -> slog::Result {
-        let Self {
-            routers_deleted,
-            routers_updated,
-            routers_created,
-            origin4_deleted,
-            origin4_updated,
-            origin4_created,
-            origin6_deleted,
-            origin6_updated,
-            origin6_created,
-            shapers_deleted,
-            shapers_updated,
-            shapers_created,
-            checkers_deleted,
-            checkers_updated,
-            checkers_created,
-            numbered_peers_deleted,
-            numbered_peers_updated,
-            numbered_peers_created,
-            unnumbered_peers_deleted,
-            unnumbered_peers_updated,
-            unnumbered_peers_created,
-        } = self;
-        for (key, val) in [
-            ("bgp-routers-deleted", routers_deleted),
-            ("bgp-routers-updated", routers_updated),
-            ("bgp-routers-created", routers_created),
-            ("bgp-origin4-deleted", origin4_deleted),
-            ("bgp-origin4-updated", origin4_updated),
-            ("bgp-origin4-created", origin4_created),
-            ("bgp-origin6-deleted", origin6_deleted),
-            ("bgp-origin6-updated", origin6_updated),
-            ("bgp-origin6-created", origin6_created),
-            ("bgp-shapers-deleted", shapers_deleted),
-            ("bgp-shapers-updated", shapers_updated),
-            ("bgp-shapers-created", shapers_created),
-            ("bgp-checkers-deleted", checkers_deleted),
-            ("bgp-checkers-updated", checkers_updated),
-            ("bgp-checkers-created", checkers_created),
-            ("bgp-numbered-peers-deleted", numbered_peers_deleted),
-            ("bgp-numbered-peers-updated", numbered_peers_updated),
-            ("bgp-numbered-peers-created", numbered_peers_created),
-            ("bgp-unnumbered-peers-deleted", unnumbered_peers_deleted),
-            ("bgp-unnumbered-peers-updated", unnumbered_peers_updated),
-            ("bgp-unnumbered-peers-created", unnumbered_peers_created),
-        ] {
-            serializer.emit_usize(key.into(), *val)?;
-        }
-        Ok(())
-    }
-}
-
-#[derive(Debug, Clone)]
-pub enum MgdBgpReconcilerStatus {
-    /// Reconciliation was skipped because we couldn't fetch the current BGP
-    /// configuration from MGD.
-    FailedReadingBgpConfig(String),
-
-    /// Reconciliation was skipped because we couldn't generate a desired config
-    /// from the bootstore.
-    ///
-    /// This should never happen - it indicates there's faulty data in the
-    /// persisted config.
-    FailedGeneratingDesiredConfig(String),
-
-    /// Reconciliation completed successfully.
-    ///
-    /// mgd operations are performed in bulk, so each item here contains the
-    /// count of items involved.
-    Success {
-        counts: MgdBgpReconcilerStatusOpCount,
-        did_change_max_paths: bool,
-    },
-
-    /// Reconciliation completed with at least one failure.
-    ///
-    /// mgd operations are performed in bulk, so each item here contains the
-    /// count of items applied on success.
-    PartialSuccess {
-        counts: MgdBgpReconcilerStatusOpCount,
-        did_change_max_paths: bool,
-        errors: Vec<String>,
-    },
-}
-
-impl slog::KV for MgdBgpReconcilerStatus {
-    fn serialize(
-        &self,
-        record: &slog::Record<'_>,
-        serializer: &mut dyn slog::Serializer,
-    ) -> slog::Result {
-        let skipped_key = "bgp-skipped".into();
-        match self {
-            Self::FailedReadingBgpConfig(reason)
-            | Self::FailedGeneratingDesiredConfig(reason) => {
-                serializer.emit_str(skipped_key, reason)
-            }
-            Self::Success { counts, did_change_max_paths } => {
-                serializer.emit_bool(
-                    "bgp-did-change-max-paths".into(),
-                    *did_change_max_paths,
-                )?;
-                counts.serialize(record, serializer)
-            }
-            MgdBgpReconcilerStatus::PartialSuccess {
-                counts,
-                did_change_max_paths,
-                errors,
-            } => {
-                serializer.emit_bool(
-                    "bgp-did-change-max-paths".into(),
-                    *did_change_max_paths,
-                )?;
-                // Each individual error is already logged; only log the count.
-                serializer.emit_usize("bgp-errors".into(), errors.len())?;
-                counts.serialize(record, serializer)
-            }
-        }
-    }
-}
 
 pub(super) async fn reconcile(
     client: &Client,
