+++ sled-agent/scrimlet-reconcilers/src/status.rs
@@ -4,80 +4,25 @@
 
 //! Types for the status and results of the reconcilers in this crate.
 
-use crate::DpdReconcilerStatus;
-use crate::LldpdReconcilerStatus;
-use crate::MgdReconcilerStatus;
-use crate::UplinkdReconcilerStatus;
+use bootstrap_agent_lockstep_types::scrimlet_reconcilers as api_status;
 use chrono::DateTime;
 use chrono::Utc;
 use std::time::Duration;
 use std::time::Instant;
 
-/// Whether or not this sled is a scrimlet.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ScrimletStatus {
-    Scrimlet,
-    NotScrimlet,
-}
+pub(crate) use api_status::DetermineSwitchSlotStatus;
+pub(crate) use api_status::ReconcilerActivationReason;
+pub(crate) use api_status::ReconcilerInertReason;
+pub(crate) use api_status::ReconciliationCompletedStatus;
+pub(crate) use api_status::ScrimletStatus;
 
-/// Status of attempting to determine this sled's switch slot via MGS within
-/// this sled's switch zone.
-#[derive(Debug, Clone)]
-pub enum DetermineSwitchSlotStatus {
-    /// We're not attempting to contact MGS because we're not a scrimlet.
-    NotScrimlet,
-
-    /// We're currently attempting to contact MGS.
-    ///
-    /// If this is not the first attempt, `prev_attempt_err` contains the error
-    /// we encountered the last time. (If the last time succeeded, we'd be
-    /// done!)
-    ContactingMgs { prev_attempt_err: Option<String> },
-
-    /// We're currently idle waiting for a timeout to retry due to a previous
-    /// failure.
-    WaitingToRetry { prev_attempt_err: String },
-}
-
-/// Why a reconciler task has gone inert.
-#[derive(Debug, Clone, Copy)]
-pub enum ReconcilerInertReason {
-    /// The reconciler task started when this sled was a scrimlet, but it has
-    /// since become "not a scrimlet" (e.g., because the attached switch has
-    /// gone away).
-    NoLongerAScrimlet,
-
-    /// The reconciler task exited. This is not expected except in tests; the
-    /// task runs forever as long as sled-agent holds on to the channels used to
-    /// communicate with it.
-    TaskExitedUnexpectedly,
-}
-
-/// Why a reconciler task was activated.
-#[derive(Debug, Clone, Copy)]
-pub enum ReconcilerActivationReason {
-    /// Each reconciler runs once on startup.
-    Startup,
-    /// The task was activated due to its periodic timer firing.
-    PeriodicTimer,
-    /// The task was activated in response to a change in the networking config.
-    SystemNetworkingConfigChanged,
-    /// The task was activated in response to the sled becoming a scrimlet again
-    /// (after previously transitioning to "not a scrimlet").
-    ScrimletStatusChanged,
-}
-
-#[derive(Debug, Clone)]
-pub struct ReconciliationCompletedStatus<T> {
-    pub activation_reason: ReconcilerActivationReason,
-    pub completed_at_time: DateTime<Utc>,
-    pub ran_for: Duration,
-    pub activation_count: u64,
-    pub status: T,
-}
+use api_status::dpd::DpdReconcilerStatus;
+use api_status::lldpd::LldpdReconcilerStatus;
+use api_status::mgd::MgdReconcilerStatus;
+use api_status::uplinkd::UplinkdReconcilerStatus;
 
 #[derive(Debug, Clone, Copy)]
-pub struct ReconcilerRunningStatus {
+pub(crate) struct ReconcilerRunningStatus {
     activation_reason: ReconcilerActivationReason,
     started_at_time: DateTime<Utc>,
     started_at_instant: Instant,
@@ -105,8 +50,18 @@ impl ReconcilerRunningStatus {
     }
 }
 
+impl From<ReconcilerRunningStatus> for api_status::ReconcilerRunningStatus {
+    fn from(value: ReconcilerRunningStatus) -> Self {
+        Self {
+            activation_reason: value.activation_reason(),
+            started_at_time: value.started_at(),
+            running_for: value.elapsed_since_start(),
+        }
+    }
+}
+
 #[derive(Debug, Clone)]
-pub enum ReconcilerCurrentStatus {
+pub(crate) enum ReconcilerCurrentStatus {
     /// The reconciler is inert: it will not or cannot run for some reason.
     Inert(ReconcilerInertReason),
     /// The reconciler is currently running.
@@ -115,18 +70,39 @@ pub enum ReconcilerCurrentStatus {
     Idle,
 }
 
-#[derive(Debug, Clone)]
-pub struct ReconcilerStatus<T> {
-    /// Status of the task at this moment.
-    pub current_status: ReconcilerCurrentStatus,
-    /// Final status of the most recent activation of this task.
-    // Box the inner status to avoid clippy complaining about
-    // `ScrimletReconcilersStatus::Running { ... }` being overly large.
-    pub last_completion: Option<Box<ReconciliationCompletedStatus<T>>>,
+impl From<ReconcilerCurrentStatus> for api_status::ReconcilerCurrentStatus {
+    fn from(value: ReconcilerCurrentStatus) -> Self {
+        match value {
+            ReconcilerCurrentStatus::Inert(reason) => Self::Inert(reason),
+            ReconcilerCurrentStatus::Running(status) => {
+                Self::Running(status.into())
+            }
+            ReconcilerCurrentStatus::Idle => Self::Idle,
+        }
+    }
 }
 
 #[derive(Debug, Clone)]
-pub enum ScrimletReconcilersStatus {
+pub(crate) struct ReconcilerStatus<T> {
+    /// Status of the task at this moment.
+    pub(crate) current_status: ReconcilerCurrentStatus,
+    /// Final status of the most recent activation of this task.
+    // Box the inner status to avoid clippy complaining about
+    // `ScrimletReconcilersStatus::Running { ... }` being overly large.
+    pub(crate) last_completion: Option<Box<ReconciliationCompletedStatus<T>>>,
+}
+
+impl<T> From<ReconcilerStatus<T>> for api_status::ReconcilerStatus<T> {
+    fn from(value: ReconcilerStatus<T>) -> Self {
+        Self {
+            current_status: value.current_status.into(),
+            last_completion: value.last_completion,
+        }
+    }
+}
+
+#[derive(Debug, Clone)]
+pub(crate) enum ScrimletReconcilersStatus {
     /// `sled-agent` has not yet provided underlay networking information.
     WaitingForSledAgentNetworkingInfo,
 
@@ -140,4 +116,28 @@ pub enum ScrimletReconcilersStatus {
         mgd_reconciler: ReconcilerStatus<MgdReconcilerStatus>,
         uplinkd_reconciler: ReconcilerStatus<UplinkdReconcilerStatus>,
     },
+}
+
+impl From<ScrimletReconcilersStatus> for api_status::ScrimletReconcilersStatus {
+    fn from(value: ScrimletReconcilersStatus) -> Self {
+        match value {
+            ScrimletReconcilersStatus::WaitingForSledAgentNetworkingInfo => {
+                Self::WaitingForSledAgentNetworkingInfo
+            }
+            ScrimletReconcilersStatus::DeterminingSwitchSlot(status) => {
+                Self::DeterminingSwitchSlot(status)
+            }
+            ScrimletReconcilersStatus::Running {
+                dpd_reconciler,
+                lldpd_reconciler,
+                mgd_reconciler,
+                uplinkd_reconciler,
+            } => Self::Running {
+                dpd_reconciler: dpd_reconciler.into(),
+                lldpd_reconciler: lldpd_reconciler.into(),
+                mgd_reconciler: mgd_reconciler.into(),
+                uplinkd_reconciler: uplinkd_reconciler.into(),
+            },
+        }
+    }
 }
